:guard: change cell kinds.
